feat: code base refactor for v1.1.0
