+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "personal-finance-for-newbies"
-version = "1.0.1"
+version = "1.1.0"
 description = ""
 authors = ["Vincenzo Ventriglia <vincenzo.ventriglia@outlook.com>"]
 readme = "README.md"
