+++ src/pages/2_📈_Return_Analysis.py
@@ -1,14 +1,10 @@
 import streamlit as st
 
-from input_output import (
-    write_disclaimer,
-    check_session_sidebar,
-    ensure_data_is_loaded,
-    get_max_common_history,
-)
-from returns import get_period_returns, get_rolling_returns
-from plot import plot_correlation_map, plot_returns, plot_rolling_returns
-from var import (
+from utils.market import get_max_common_history
+from core.returns import get_period_returns, get_rolling_returns
+from utils.ui import write_disclaimer, check_session_sidebar, ensure_data_is_loaded
+from utils.plot import plot_correlation_map, plot_returns, plot_rolling_returns
+from utils.var import (
     GLOBAL_STREAMLIT_STYLE,
     PLT_CONFIG_NO_LOGO,
     FAVICON,
